Added multiple functions for the board
added is_king to track if a piece is a king, simple as. valid_position checks for valid positions on the board. evaluate_board is a setup function to help the AI determine if a piece is worthy to be taken. also added check_winner which does nothing right now but check for winners while testing