The ginormous minimax update
finally got around to this working, took a lot of concepts from another lab in order to create the minimax function. Since a player is actually playing this time I need to move most move functions out of the agent and instead separated them to be accessible by the player. I also took an excruciating amount of time trying to minimize the amount of functions needed to work inside of play_game. I am going to work next on trying to simplify how to move in the game to be more tied to battleship instead of staring at row and column numbers for the entire session.